+++ GesdinetJWTRefreshTokenBundle.php
@@ -6,7 +6,6 @@ use Gesdinet\JWTRefreshTokenBundle\DependencyInjection\Compiler\AddExtractorsToC
 use Gesdinet\JWTRefreshTokenBundle\DependencyInjection\Compiler\CustomUserProviderCompilerPass;
 use Gesdinet\JWTRefreshTokenBundle\DependencyInjection\Compiler\ObjectManagerCompilerPass;
 use Gesdinet\JWTRefreshTokenBundle\DependencyInjection\Compiler\UserCheckerCompilerPass;
-use Gesdinet\JWTRefreshTokenBundle\DependencyInjection\Security\Factory\LegacyRefreshTokenAuthenticatorFactory;
 use Gesdinet\JWTRefreshTokenBundle\DependencyInjection\Security\Factory\RefreshTokenAuthenticatorFactory;
 use Symfony\Bundle\SecurityBundle\DependencyInjection\SecurityExtension;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -24,16 +23,11 @@ class GesdinetJWTRefreshTokenBundle extends Bundle
         $container->addCompilerPass(new ObjectManagerCompilerPass());
         $container->addCompilerPass(new UserCheckerCompilerPass(true));
 
-        // Only register the security authenticator for Symfony 5.3+
+        // Only register the security authenticator for Symfony 5.4+
         if (interface_exists(RememberMeHandlerInterface::class)) {
             /** @var SecurityExtension $extension */
             $extension = $container->getExtension('security');
-
-            if (method_exists($extension, 'addAuthenticatorFactory')) {
-                $extension->addAuthenticatorFactory(new RefreshTokenAuthenticatorFactory());
-            } else {
-                $extension->addSecurityListenerFactory(new LegacyRefreshTokenAuthenticatorFactory());
-            }
+            $extension->addAuthenticatorFactory(new RefreshTokenAuthenticatorFactory());
         }
     }
 }