+++ Tests/Functional/ORMTestCase.php
@@ -27,8 +27,18 @@ abstract class ORMTestCase extends TestCase
             $config->setMetadataCacheImpl(DoctrineProvider::wrap(new ArrayAdapter()));
         }
 
-        $config->setQueryCacheImpl(DoctrineProvider::wrap(new ArrayAdapter()));
-        $config->setResultCacheImpl(DoctrineProvider::wrap(new ArrayAdapter()));
+        if (method_exists($config, 'setQueryCache')) {
+            $config->setQueryCache(new ArrayAdapter());
+        } else {
+            $config->setQueryCacheImpl(DoctrineProvider::wrap(new ArrayAdapter()));
+        }
+
+        if (method_exists($config, 'setResultCache')) {
+            $config->setResultCache(new ArrayAdapter());
+        } else {
+            $config->setResultCacheImpl(DoctrineProvider::wrap(new ArrayAdapter()));
+        }
+
         $config->setProxyDir(sys_get_temp_dir().'/JWTRefreshTokenBundle/_files');
         $config->setProxyNamespace(__NAMESPACE__.'\Proxies');